ajustando o componente da pagina inicial, criando um arquivo para guardar os metodos
